tquic: fix security and correctness issues from review
Address critical, high, and medium severity issues found during
comprehensive review of the security hardening commit:

Crypto key update fixes:
- Fix key material leak on stack when first HKDF call fails by
  changing early 'return err' to 'goto out' so memzero_explicit
  runs in quic_crypto.c and quic_key_update.c (tx and rx paths)
- Defer secret commit until after all derived values (key, IV, AEAD)
  are successfully installed, preventing inconsistent crypto state
  on partial failure

NAPI crash prevention:
- Add minimal ndo_start_xmit handler to dummy net_device_ops that
  drops stray packets, preventing NULL function pointer panic if
  a packet is ever queued to the NAPI dummy device

Crypto resource management:
- Free all crypto handles (AEAD, skcipher, shash) in
  tquic_crypto_wrapper_free before zeroizing and freeing, preventing
  kernel crypto handle leaks on connection teardown
- Add HKDF label bounds checks to handshake.c and tls.c parallel
  implementations matching the check already in quic_crypto.c

Concurrency fixes:
- Fix retry token AEAD use-after-free by keeping spinlock held
  across crypto_aead_setkey and aead_request_alloc calls, preventing
  concurrent tquic_retry_state_free from freeing the AEAD transform
- Fix migration path recovery use-after-free by saving path_id under
  lock and re-looking up the path after releasing it, instead of
  dereferencing a potentially stale pointer

Other fixes:
- Fix integer overflow in rate calculation: cast to u64 before
  multiplying count by 1000 to prevent signed overflow under attack
- Fix trailing whitespace in quic_key_update.c and quic_socket.c

+++ net/tquic/tquic_ratelimit.c
@@ -948,7 +948,7 @@ static void tquic_rl_rate_calc_work_fn(struct work_struct *work)
 		elapsed_ms = 1;
 
 	count = atomic_xchg(&state->rate_window_count, 0);
-	rate = (count * 1000) / elapsed_ms;
+	rate = (int)(((u64)count * 1000) / elapsed_ms);
 
 	state->rate_window_start = now;
